chore(facade): use formula facade test bed
